requirements.txt
google-generativeai
python-slugify
python-dotenv
